+++ CHANGELOG.md
@@ -15,6 +15,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 #### :bug: Bug Fix
 
+* Fixed soft transitions with array parameters `base/b-router`
 * Fixed an issue with testing userAgent string for desktop Safari `core/browser`
 
 ## v3.30.1 (2022-10-25)